fix: minor compilation issues

+++ includes/time.h
@@ -2,6 +2,7 @@
 #define TIME_H
 
 #include "standard.h"
+#include "bcm2835.h"
 
 #define ST_REG(a) (BCM2835_ST + (uint32_t)(a))
 
@@ -12,7 +13,7 @@ typedef enum e_st_offset
 	ST_CHI = 0x08u
 }	t_st_offset;
 
-uint32_t get_time_us(void);
+uint64_t get_time_us(void);
 uint32_t get_time_ms(void);
 void usleep(uint32_t usec);
 void msleep(uint32_t msec);
